+++ Tests/Test-CustomObject.ps1
@@ -13,10 +13,11 @@ $custom = Get-Process mongod | Select-Object $property -First 1
 $document = New-MdbcData $custom
 if ($document.GetType().FullName -ne 'Mdbc.Dictionary') { throw }
 
-### Many custom objects as input objects
-# ??? make use of -DocumentId and -GenerateNewId in such a scenario
-# ??? think to simplify [New-Object psobject -Property $_] -- say, Format-Mdbc
-Get-Process svchost | Select-Object $property | New-MdbcData | %{
-	if ($_.GetType().FullName -ne 'Mdbc.Dictionary') { throw }
-	New-Object psobject -Property $_
-}
+### Many custom objects from pipeline
+$documents = Get-Process svchost | Select-Object $property | New-MdbcData
+foreach($_ in $documents) {	if ($_.GetType().FullName -ne 'Mdbc.Dictionary') { throw } }
+
+### Convert back to custom objects with a helper and manually
+$custom1 = $documents | Convert-MdbcData
+$custom2 = $documents | %{ New-Object psobject -Property $_ }
+$custom1 | Format-Table -AutoSize | Out-String